refactor load block
